+++ CHANGES.md
@@ -1,3 +1,10 @@
+## 1.2.0
+  * Removed Image and draw
+  * Added DrawMode and DrawRoute
+  * Added DrawT and runDrawT, removed runDraw
+  * Renamed mapVertices to mapGeometry
+  * Added foldMapGeometry and autoElements
+
 ## 1.1.0
 
   * Renamed (&#126;&#126;) to (&#126;&lt;) and (&#126;&#42;) to (&#126;&lt;&#42;)